+++ beyond/frames/frames.py
@@ -191,8 +191,7 @@ def orbit2frame(name, ref_orbit, orientation=None, parent=EME2000, exists_warnin
         ref_orbit (Orbit or Ephem):
         orientation (str): Orientation of the created frame. If orientation is ``None``,
             the new frame will keep the orientation of the reference frame of the Orbit
-            and move along with the orbit. Other acceptable values are ``"QSW"`` (and its
-            aliases "LVLH" and "RSW") or ``"TNW"``.
+            and move along with the orbit. Other acceptable values are ``"QSW"`` or ``"TNW"``.
         parent (Frame) : Inertial frame to which to attach the LOF
         exists_warning (bool): Disable the warning when creating a frame with an already
             taken name by setting this value to False
@@ -205,9 +204,7 @@ def orbit2frame(name, ref_orbit, orientation=None, parent=EME2000, exists_warnin
     if orientation is None:
         orientation = ref_orbit.frame.orientation
     else:
-        if orientation.upper() in ("RSW", "LVLH"):
-            orientation = "QSW"
-        elif orientation.upper() not in ("QSW", "TNW"):
+        if orientation.upper() not in ("QSW", "TNW"):
             raise ValueError("Unknown orientation '%s'" % orientation)
 
         orientation = orient.LocalOrbitalOrientation(