+++ beyond/orbits/statevector.py
@@ -191,7 +191,7 @@ StateVector =
     def cov(self, value):
 
         if not isinstance(value, Cov):
-            raise TypeError("Unknwon covariance type : ".format(type(value)))
+            raise TypeError("Unknwon covariance type : {}".format(type(value)))
 
         self._data["cov"] = value
         self._data["cov"].orb = self
@@ -235,8 +235,8 @@ StateVector =
 
             orbit.form = "cartesian"
             # is equivalent to
-            from beyond.orbits.forms import FormTransform
-            orbit.form = FormTransform.CART
+            from beyond.orbits.forms import CART
+            orbit.form = CART
         """
         return self._data["form"]