+++ tests/orbits/test_forms.py
@@ -3,8 +3,8 @@ from pytest import fixture
 import numpy as np
 
 from beyond.dates import Date
-from beyond.orbits import Orbit
-from beyond.orbits.forms import TLE, KEPL_E, KEPL_C, KEPL_M, KEPL, SPHE, CART
+from beyond.orbits import StateVector
+from beyond.orbits.forms import TLE, KEPL_E, KEPL_C, KEPL_M, KEPL, SPHE, CART, EQUI, CYL, KEPL_MC
 
 
 @fixture
@@ -14,98 +14,131 @@ def ref_date():
 
 @fixture
 def ref_kepl_m(ref_date):
-    ref_coord = [
-        7192631.11295, 0.00218439, np.deg2rad(98.50639978),
-        np.deg2rad(315.72521309), np.deg2rad(67.54633938), np.deg2rad(174.37062038)
-    ]
-    ref_form = KEPL_M
-    ref_frame = "EME2000"
-    ref_propagator = 'Sgp4'
-    return Orbit(ref_coord, ref_date, ref_form, ref_frame, ref_propagator)
-
+    return StateVector(
+        [
+            7192631.11295, 0.00218439, np.deg2rad(98.50639978),
+            np.deg2rad(315.72521309), np.deg2rad(67.54633938), np.deg2rad(174.37062038)
+        ],
+        ref_date,
+        KEPL_M,
+        "EME2000"
+    )
 
 @fixture
 def ref_kepl(ref_date):
 
-    return Orbit(
+    return StateVector(
         [
             7192631.11295, 0.00218439, 1.71926101045,
             5.51044449998, 1.17890601985, 3.04376883055,
         ],
         ref_date,
         KEPL,
-        "EME2000",
-        "Sgp4"
+        "EME2000"
     )
 
 
 @fixture
 def ref_cart(ref_date):
-    return Orbit(
+    return StateVector(
         [
             -1.77079285e+06, 3.04066799e+06, -6.29108469e+06,
             5.05386814e+03, -4.20539932e+03, -3.45695948e+03
         ],
         ref_date,
         CART,
-        "EME2000",
-        "Sgp4"
+        "EME2000"
     )
 
 
 @fixture
 def ref_sphe(ref_date):
-    return Orbit(
+    return StateVector(
         [
             7.20826718e+06, 2.09815148e+00, -1.06082733e+00,
             1.58820333e+00, -6.39690485e-04, -9.82054932e-04
         ],
         ref_date,
         SPHE,
-        "EME2000",
-        "Sgp4"
+        "EME2000"
     )
 
 
 @fixture
 def ref_kepl_c(ref_date):
-    return Orbit(
+    return StateVector(
         [
             7.19263111e+06, 8.34297391e-04, 2.01878863e-03,
             1.71926101e+00, 5.51044450e+00, 4.22224746e+00
         ],
         ref_date,
         KEPL_C,
-        "EME2000",
-        "Sgp4"
+        "EME2000"
     )
 
 
 @fixture
 def ref_tle(ref_date):
-    return Orbit(
+    return StateVector(
         [
             1.71926101e+00, 5.51044450e+00, 2.18439000e-03,
             1.17890602e+00, 3.04334144e+00, 1.03499315e-03
         ],
         ref_date,
         TLE,
-        "EME2000",
-        "Sgp4"
+        "EME2000"
     )
 
 
 @fixture
 def ref_kepl_hyper(ref_date):
-    return Orbit(
+    return StateVector(
         [
             -7192631.11295, 1.00218439, 1.71926101045,
             5.51044449998, 1.17890601985, 0.1,
         ],
         ref_date,
         KEPL,
-        "EME2000",
-        None
+        "EME2000"
+    )
+
+
+@fixture
+def ref_equi(ref_date):
+    return StateVector(
+        [
+            7.19263111e+06,  2.00667384e-03,  8.63029427e-04,
+            8.31052761e-01, -8.10276759e-01,  9.73311935e+00
+        ],
+        ref_date,
+        EQUI,
+        "EME2000"
+    )
+
+
+@fixture
+def ref_cyl(ref_date):
+    return StateVector(
+        [
+            3.51871697e+06,  2.09815148e+00, -6.29108469e+06,
+            -6.17741547e+03, -6.39690485e-04, -3.45695948e+03
+        ],
+        ref_date,
+        CYL,
+        "EME2000"
+    )
+
+
+@fixture
+def ref_kepl_mc(ref_date):
+    return StateVector(
+        [
+            7.19263111e+06, 8.34297391e-04, 2.01878863e-03,
+            1.71926101e+00, 5.51044450e+00, 4.22224746e+00
+        ],
+        ref_date,
+        KEPL_MC,
+        "EME2000"
     )
 
 
@@ -225,17 +258,16 @@ def test_tle(ref_tle):
 
 def test_hyperbolic(ref_kepl_hyper):
 
-    def to_orbit(arr, form):
-        return Orbit(
+    def to_sv(arr, form):
+        return StateVector(
             arr,
             ref_kepl_hyper.date,
             form,
             ref_kepl_hyper.frame,
-            None
         )
 
     # To cartesian ...
-    cart = to_orbit(KEPL(ref_kepl_hyper, CART), CART.name)
+    cart = to_sv(KEPL(ref_kepl_hyper, CART), CART.name)
     assert np.allclose(cart, [
         1687.60352732, -4761.78462622, 14918.72031867,
         -159632.45044557, 140048.25302776, 74649.10622557
@@ -246,23 +278,50 @@ def test_hyperbolic(ref_kepl_hyper):
     assert np.allclose(kepl_h, ref_kepl_hyper)
 
     # To keplerian eccentric ...
-    kepl_e = to_orbit(KEPL(ref_kepl_hyper, KEPL_E), KEPL_E.name)
+    kepl_e = to_sv(KEPL(ref_kepl_hyper, KEPL_E), KEPL_E.name)
     assert np.allclose(kepl_e, [
         -7.19263111e+06, 1.00218439e+00, 1.71926101e+00,
         5.51044450e+00, 1.17890602e+00, 3.30579237e-03
     ])
 
     # ... and back
-    kepl_h = to_orbit(KEPL_E(kepl_e, KEPL), KEPL.name)
+    kepl_h = to_sv(KEPL_E(kepl_e, KEPL), KEPL.name)
     assert np.allclose(kepl_h, ref_kepl_hyper)
 
     # To Keplerian mean ...
-    kepl_m = to_orbit(KEPL_E(kepl_e, KEPL_M), KEPL_M.name)
+    kepl_m = to_sv(KEPL_E(kepl_e, KEPL_M), KEPL_M.name)
     assert np.allclose(kepl_m, [
         -7.19263111e+06, 1.00218439e+00, 1.71926101e+00,
         5.51044450e+00, 1.17890602e+00, 7.22717405e-06
     ])
 
     # ... and back
-    kepl_e2 = to_orbit(KEPL_M(kepl_m, KEPL_E), KEPL_E.name)
+    kepl_e2 = to_sv(KEPL_M(kepl_m, KEPL_E), KEPL_E.name)
     assert np.allclose(kepl_e, kepl_e2)
+
+
+def test_equinoctial(ref_kepl, ref_equi):
+
+    equi = KEPL(ref_kepl, EQUI)
+    assert np.allclose(ref_equi, equi)
+
+    kepl = EQUI(ref_equi, KEPL)
+    assert np.allclose(ref_kepl, kepl)
+
+
+def test_cylindrical(ref_cart, ref_cyl):
+
+    cyl = CART(ref_cart, CYL)
+    assert np.allclose(ref_cyl, cyl)
+
+    cart = CYL(ref_cyl, CART)
+    assert np.allclose(ref_cart, cart)
+
+
+def test_keplerian_mean_circular(ref_kepl_m, ref_kepl_mc):
+
+    kepl_mc = KEPL_M(ref_kepl_m, KEPL_MC)
+    assert np.allclose(ref_kepl_mc, kepl_mc)
+
+    kepl_m = KEPL_MC(ref_kepl_mc, KEPL_M)
+    assert np.allclose(ref_kepl_m, kepl_m)