+++ CHANGELOG.md
@@ -3,6 +3,19 @@
 This file tries to regroup all notable modifications of the ``beyond`` library.
 Each release is linked to a git commit.
 
+## [Unreleased]
+
+### Added
+
+- equinoctial, cylindrical and mean circular forms
+- OMM in SGP and TLE formats are compatibles
+- A lot of small additions to the documentation of existing functions/methods
+
+### Removed
+
+- RSW and LVLV are not synonym to QSW any more
+- Disable keplerian representation of non inertial CCSDS OPM
+
 ## [v0.7.1] - 2020-09-13
 
 ### Added